Version 0.1
The first version of my god-awful website. Wow, this thing is ugly.

But...it works! I did it! I built a website completely from scratch, with images!

I intend to allow users to click through different versions of this website to see my progress and learning over time. As I gain an understanding of the tools available, I will periodically update this page.

This is not a portfolio site!

It's an ugly, raw, poorly designed learning tool. 

And I'm gosh darn proud of it.